fixing serialization to only write to bundle.ml transformers that are supported by spark

+++ python/mleap/sklearn/pipeline.py
@@ -15,19 +15,16 @@
 # limitations under the License.
 #
 
-import sklearn
 from sklearn.pipeline import Pipeline
-from sklearn.pipeline import FeatureUnion
 import os
 import json
 import shutil
 import uuid
 
 
-
-def serialize_to_bundle(self, path, model_name):
+def serialize_to_bundle(self, path, model_name, init=False):
     serializer = SimpleSparkSerializer()
-    serializer.serialize_to_bundle(self, path, model_name)
+    serializer.serialize_to_bundle(self, path, model_name, init)
 
 
 def deserialize_from_bundle(self, path):
@@ -38,36 +35,49 @@ setattr(Pipeline, 'serialize_to_bundle', serialize_to_bundle)
 setattr(Pipeline, 'deserialize_from_bundle', deserialize_from_bundle)
 setattr(Pipeline, 'op', 'pipeline')
 setattr(Pipeline, 'name', "{}_{}".format('pipeline', uuid.uuid1()))
+setattr(Pipeline, 'serializable', True)
 
 
 class SimpleSparkSerializer(object):
     def __init__(self):
         super(SimpleSparkSerializer, self).__init__()
 
-    def serialize_to_bundle(self, transformer, path, model_name):
+    def serialize_to_bundle(self, transformer, path, model_name, init=False):
 
-        if os.path.exists("{}/{}".format(path, model_name)):
-            shutil.rmtree("{}/{}".format(path, model_name))
+        model_dir = path
+        if init:
+            # If bundle path already exists, delte it and create a clean directory
+            if os.path.exists("{}/{}".format(path, model_name)):
+                shutil.rmtree("{}/{}".format(path, model_name))
 
-        model_dir = "{}/{}".format(path, model_name)
-        os.mkdir(model_dir)
+            model_dir = "{}/{}".format(path, model_name)
+            os.mkdir(model_dir)
 
-        for step in [x[1] for x in transformer.steps]:
+        # Write Pipeline Bundle file
+        with open("{}/{}".format(model_dir, 'bundle.json'), 'w') as outfile:
+            json.dump(self.get_bundle(transformer), outfile, indent=3)
+
+        for step in [x[1] for x in transformer.steps if hasattr(x[1], 'serialize_to_bundle')]:
             name = step.name
             print(name)
-            bundle_dir = "{}/{}/{}.node".format(path, model_name, name)
-            os.mkdir(bundle_dir)
+
             if step.op == 'pipeline':
-                # Write Model and Node .json files
-                with open("{}/{}".format(bundle_dir, 'model.json'), 'w') as outfile:
-                    json.dump(step.get_mleap_model(), outfile, indent=3)
-                with open("{}/{}".format(bundle_dir, 'node.json'), 'w') as f:
-                    json.dump(step.get_mleap_node(), f, indent=3)
+                # Create the node directory
+                bundle_dir = "{}/{}.node".format(model_dir, name)
+                os.mkdir(bundle_dir)
+
+                # Write bundle file
+                with open("{}/{}".format(bundle_dir, 'bundle.json'), 'w') as outfile:
+                    json.dump(self.get_bundle(step), outfile, indent=3)
 
                 for step_i in [x[1] for x in step.steps]:
                     step_i.serialize_to_bundle(bundle_dir)
+            elif step.op == 'feature_union':
+                print type(step)
+                print dir(step.serialize_to_bundle)
+                step.serialize_to_bundle(model_dir, step.name)
             else:
-                step.serialize_to_bundle(bundle_dir)
+                step.serialize_to_bundle(model_dir, step.name)
 
             if isinstance(step, list):
                 pass
@@ -80,7 +90,16 @@ class SimpleSparkSerializer(object):
           "name": transformer.name,
           "format": "json",
           "version": "0.4.0-SNAPSHOT",
-          "nodes": [x.name for x in transformer.steps]
+          "nodes": self._extract_nodes(transformer.steps)
         }
         return js
 
+    def _extract_nodes(self, steps):
+        pipeline_steps = []
+        for name, step in steps:
+            if step.op == 'feature_union':
+                union_steps = [x[1].name for x in step.transformer_list if hasattr(x[1], 'serialize_to_bundle')]
+                pipeline_steps += union_steps
+            elif hasattr(step, 'serialize_to_bundle'):
+                pipeline_steps.append(name)
+        return pipeline_steps
